requirements.txt
falcon==2.0.0
google-cloud-firestore==1.6.2
gunicorn==20.0.4
pydialogflow-fulfillment==0.1.3
pymongo==3.10.1
pyyaml==5.3.1